Harden share card privacy and naming
